+++ src/Telegram.Bot/Types/Message.cs
@@ -72,6 +72,12 @@ namespace Telegram.Bot.Types
         public string ForwardSignature { get; set; }
 
         /// <summary>
+        /// Optional. Sender's name for messages forwarded from users who disallow adding a link to their account in forwarded messages
+        /// </summary>
+        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
+        public string ForwardSenderName { get; set; }
+
+        /// <summary>
         /// Optional. For forwarded messages, date the original message was sent in Unix time
         /// </summary>
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
@@ -217,6 +223,12 @@ namespace Telegram.Bot.Types
         /// </summary>
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
         public Venue Venue { get; set; }
+
+        /// <summary>
+        /// Optional. Message is a native poll, information about the poll
+        /// </summary>
+        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
+        public Poll Poll { get; set; }
 
         /// <summary>
         /// Optional. New members that were added to the group or supergroup and information about them (the bot itself may be one of these members)
@@ -395,6 +407,9 @@ namespace Telegram.Bot.Types
 
                 if (MigrateToChatId != default)
                     return MessageType.MigratedToSupergroup;
+
+                if (Poll != null)
+                    return MessageType.Poll;
 
                 return MessageType.Unknown;
             }