+++ test/Telegram.Bot.Tests.Integ/Framework/TestCollectionOrderer.cs
@@ -6,6 +6,7 @@ using Xunit.Abstractions;
 
 namespace Telegram.Bot.Tests.Integ.Framework
 {
+    // ReSharper disable once UnusedMember.Global
     public class TestCollectionOrderer : ITestCollectionOrderer
     {
         private readonly string[] _orderedCollections =
@@ -23,6 +24,7 @@ namespace Telegram.Bot.Tests.Integ.Framework
             Constants.TestCollections.Payment,
             Constants.TestCollections.ChatMemberAdministration,
             Constants.TestCollections.Exceptions,
+            Constants.TestCollections.NativePolls,
 
             // Tests without the need for user interaction:
             Constants.TestCollections.GettingUpdates,