+++ test/Telegram.Bot.Tests.Integ/Sending Messages/AudioMessageTests.cs
@@ -24,15 +24,13 @@ namespace Telegram.Bot.Tests.Integ.Sending_Messages
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendAudio)]
         public async Task Should_Send_Audio()
         {
-            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldSendAudio);
-
             const string performer = "Jackson F. Smith";
             const string title = "Cantina Rag";
             const int duration = 201;
             const string caption = "Audio File in .mp3 format";
 
             Message message;
-            using (Stream stream = System.IO.File.OpenRead(Constants.FileNames.Audio.CantinaRagMp3))
+            using (Stream stream = System.IO.File.OpenRead(Constants.PathToFile.Audio.CantinaRagMp3))
             {
                 message = await BotClient.SendAudioAsync(
                     chatId: _fixture.SupergroupChat,
@@ -58,13 +56,11 @@ namespace Telegram.Bot.Tests.Integ.Sending_Messages
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendAudio)]
         public async Task Should_Send_Audio_With_Thumb()
         {
-            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldSendAudioWithThumbnail);
-
             // Both audio file and its thumbnail should be uploaded
             Message message;
             using (Stream
-                stream1 = System.IO.File.OpenRead(Constants.FileNames.Audio.AStateOfDespairMp3),
-                stream2 = System.IO.File.OpenRead(Constants.FileNames.Thumbnail.TheAbilityToBreak)
+                stream1 = System.IO.File.OpenRead(Constants.PathToFile.Audio.AStateOfDespairMp3),
+                stream2 = System.IO.File.OpenRead(Constants.PathToFile.Thumbnail.TheAbilityToBreak)
             )
             {
                 message = await BotClient.SendAudioAsync(
@@ -85,13 +81,11 @@ namespace Telegram.Bot.Tests.Integ.Sending_Messages
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendVoice)]
         public async Task Should_Send_Voice()
         {
-            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldSendVoice);
-
             const int duration = 24;
             const string caption = "Test Voice in .ogg format";
 
             Message message;
-            using (var stream = System.IO.File.OpenRead(Constants.FileNames.Audio.TestOgg))
+            using (var stream = System.IO.File.OpenRead(Constants.PathToFile.Audio.TestOgg))
             {
                 message = await BotClient.SendVoiceAsync(
                     chatId: _fixture.SupergroupChat,